requirements.txt
flask==2.0.1
werkzeug==2.0.1
nltk==3.6.2
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0 